+++ include/databento/enums.hpp
@@ -48,6 +48,15 @@ enum class DatasetCondition : std::uint8_t {
   Missing,
 };
 
+// Live session parameter which controls gateway behavior when the client
+// falls behind real time.
+enum class SlowReadBehavior : std::uint8_t {
+  // Send a warning but continue reading.
+  Warn = 0,
+  // Skip records to catch up.
+  Skip = 1,
+};
+
 // A record type sentinel.
 namespace r_type {
 enum RType : std::uint8_t {
@@ -662,6 +671,7 @@ const char* ToString(SplitDuration duration_interval);
 const char* ToString(Delivery delivery);
 const char* ToString(JobState state);
 const char* ToString(DatasetCondition condition);
+const char* ToString(SlowReadBehavior slow_read_behavior);
 const char* ToString(RType r_type);
 const char* ToString(Side side);
 const char* ToString(Action action);
@@ -688,6 +698,7 @@ std::ostream& operator<<(std::ostream& out, SplitDuration duration_interval);
 std::ostream& operator<<(std::ostream& out, Delivery delivery);
 std::ostream& operator<<(std::ostream& out, JobState state);
 std::ostream& operator<<(std::ostream& out, DatasetCondition condition);
+std::ostream& operator<<(std::ostream& out, SlowReadBehavior slow_read_behavior);
 std::ostream& operator<<(std::ostream& out, RType r_type);
 std::ostream& operator<<(std::ostream& out, Side side);
 std::ostream& operator<<(std::ostream& out, Action action);