+++ include/databento/live_threaded.hpp
@@ -55,6 +55,7 @@ class LiveThreaded {
   VersionUpgradePolicy UpgradePolicy() const;
   std::optional<std::chrono::seconds> HeartbeatInterval() const;
   databento::Compression Compression() const;
+  std::optional<databento::SlowReadBehavior> SlowReadBehavior() const;
   const std::vector<LiveSubscription>& Subscriptions() const;
   std::vector<LiveSubscription>& Subscriptions();
 
@@ -109,13 +110,15 @@ class LiveThreaded {
                bool send_ts_out, VersionUpgradePolicy upgrade_policy,
                std::optional<std::chrono::seconds> heartbeat_interval,
                std::size_t buffer_size, std::string user_agent_ext,
-               databento::Compression compression);
+               databento::Compression compression,
+               std::optional<databento::SlowReadBehavior> slow_read_behavior);
   LiveThreaded(ILogReceiver* log_receiver, std::string key, std::string dataset,
                std::string gateway, std::uint16_t port, bool send_ts_out,
                VersionUpgradePolicy upgrade_policy,
                std::optional<std::chrono::seconds> heartbeat_interval,
                std::size_t buffer_size, std::string user_agent_ext,
-               databento::Compression compression);
+               databento::Compression compression,
+               std::optional<databento::SlowReadBehavior> slow_read_behavior);
 
   // unique_ptr to be movable
   std::unique_ptr<Impl> impl_;